refactor(desktop): remove `app_search_catalog` in favor of `cached`
