+++ harper-desktop/src-tauri/src/mac_broker/app_catalog.rs
@@ -28,7 +28,8 @@ fn display_name_from_app_path(path: &str) -> Option<String> {
     }
 }
 
-pub fn installed_application_bundle_ids() -> Result<Vec<String>, String> {
+#[cached]
+pub fn installed_application_bundle_ids() -> Result<Arc<Vec<String>>, String> {
     let output = Command::new("mdfind")
         .arg(format!(
             "kMDItemContentType == \"{APPLICATION_BUNDLE_CONTENT_TYPE}\""
@@ -45,7 +46,7 @@ pub fn installed_application_bundle_ids() -> Result<Vec<String>, String> {
         .filter_map(|line| bundle_id_from_app_path(line.trim()))
         .collect::<Vec<_>>();
 
-    Ok(deduplicate_and_sort_bundle_ids(bundle_ids))
+    Ok(Arc::new(deduplicate_and_sort_bundle_ids(bundle_ids)))
 }
 
 /// Resolves a bundle identifier to an installed `.app` path using Spotlight metadata.
